+++ digest/tarsum_test.go
@@ -1,4 +1,4 @@
-package common
+package digest
 
 import (
 	"reflect"
